Added Main Menu
Added Main menu that allows the player to bounce Bloopy while they are on the menu. Also added the beignings of the new Launch system. Will have to implament launching with a launch object soon. Plan on  using annimation when start  button is pressed to move Bloopy to the launch object.