Bllopy Bounce Fixed and Camera Bounds Added
Bloopy's bouncing has been fixed. Issue was in BloopyBehaviors, when hScreenRange was set to 0, Bloopy position was constantly being reset to 0. I also added checks and calculations in BloopyBehavior.Start() to set hScreenRange and vScreenRange based on the size of the camera, which should allow for these to work for differing screen sizes. Testing on external devices required.