Binary Save File System Added
High scores are now being saved wit a Binary Save File System. Fairly simple for far, but with room for additional stuff to be added. SaveSystem checks if a save file exists and creates on if it doesn't on Awake(), which is done on the main menu. This should ensure there is always a save file for the game to reference.